Big update (#1021)
* fix typo

* fix connection error because of api version

* RDM002

* rdm002

* update mqtt sensor

* fix sensor type

* add remove apiuser

* deprecated utcnow

The method "utcnow" in class "datetime" is deprecated
  Use timezone-aware objects to represent datetimes in UTC; e.g. by calling .now(datetime.UTC)

* Update __init__.py

* Update configHandler.py

* Update views.py

* add config setting

* Update configHandler.py

* Update daylightSensor.py

* Update stateFetch.py

* Update stateFetch.py

* Update stateFetch.py

* Update stateFetch.py

* Update daylightSensor.py

* Update discover.py

* Update scheduler.py

* replace localhost with IP

* update views.py, githubinstall.sh, updateManager.py

* download debug

* Update configHandler.py

* Update configHandler.py

* Update configHandler.py

* Update configHandler.py

* Update configHandler.py

* Update configHandler.py

* Update configHandler.py

* Update configHandler.py

* Update logger.py

* Update configHandler.py

* Update configHandler.py

* Update configHandler.py

* Update configHandler.py

* Update configHandler.py

* Update bug-report.yml

* Update bug-report.yml

* log

* Update bug-report.yml

* Update logger.py

* remove sleep, add WebUI version, fix logfile loglevel

* update installer, logger, bug-report

* delete old webui

* more remove old ui

* Update bug-report.yml

* Update Dockerfile

* move diyhue.log

+++ BridgeEmulator/githubInstall.sh
@@ -1,38 +1,37 @@
-curl -s localhost/save
+curl -s $1/save
 cd /
-#curl -s -J -L -o diyhue.tar.gz "https://github.com/diyhue/diyHue/archive/refs/heads/master.tar.gz"
-#curl -s -J -L -o diyhue.tar.gz "https://github.com/hendriksen-mark/diyhue/archive/refs/heads/master.tar.gz"
-curl -sL -o diyhue.zip https://github.com/diyhue/diyhue/archive/master.zip
-#curl -sL -o diyhue.zip https://github.com/hendriksen-mark/diyhue/archive/master.zip
-#tar xzf diyhue.tar.gz --strip-components=1 -C diyhue
-unzip -qo diyhue.zip
-rm diyhue.zip
-#cd diyhue
-cp -r diyHue-master/BridgeEmulator/flaskUI /opt/hue-emulator/
-cp -r diyHue-master/BridgeEmulator/functions /opt/hue-emulator/
-cp -r diyHue-master/BridgeEmulator/lights /opt/hue-emulator/
-cp -r diyHue-master/BridgeEmulator/sensors /opt/hue-emulator/
-cp -r diyHue-master/BridgeEmulator/HueObjects /opt/hue-emulator/
-cp -r diyHue-master/BridgeEmulator/services /opt/hue-emulator/
-cp -r diyHue-master/BridgeEmulator/configManager /opt/hue-emulator/
-cp -r diyHue-master/BridgeEmulator/logManager /opt/hue-emulator/
-cp -r diyHue-master/BridgeEmulator/HueEmulator3.py /opt/hue-emulator/
-cp -r diyHue-master/BridgeEmulator/githubInstall.sh /opt/hue-emulator/
-cp -r diyHue-master/BridgeEmulator/githubUIInstall.sh /opt/hue-emulator/
-cp -r diyHue-master/BridgeEmulator/genCert.sh /opt/hue-emulator/
-cp -r diyHue-master/BridgeEmulator/openssl.conf /opt/hue-emulator/
-chmod +x /opt/hue-emulator/genCert.sh
+if [ $2 = allreadytoinstall ]; then
+    echo "diyhue + ui update"
+    curl -sL -o diyhue.zip https://github.com/diyhue/diyhue/archive/master.zip
+    #curl -sL -o diyhue.zip https://github.com/hendriksen-mark/diyhue/archive/master.zip
+    unzip -qo diyhue.zip
+    rm diyhue.zip
+    cp -r diyHue-master/BridgeEmulator/flaskUI /opt/hue-emulator/
+    cp -r diyHue-master/BridgeEmulator/functions /opt/hue-emulator/
+    cp -r diyHue-master/BridgeEmulator/lights /opt/hue-emulator/
+    cp -r diyHue-master/BridgeEmulator/sensors /opt/hue-emulator/
+    cp -r diyHue-master/BridgeEmulator/HueObjects /opt/hue-emulator/
+    cp -r diyHue-master/BridgeEmulator/services /opt/hue-emulator/
+    cp -r diyHue-master/BridgeEmulator/configManager /opt/hue-emulator/
+    cp -r diyHue-master/BridgeEmulator/logManager /opt/hue-emulator/
+    cp -r diyHue-master/BridgeEmulator/HueEmulator3.py /opt/hue-emulator/
+    cp -r diyHue-master/BridgeEmulator/githubInstall.sh /opt/hue-emulator/
+    cp -r diyHue-master/BridgeEmulator/genCert.sh /opt/hue-emulator/
+    cp -r diyHue-master/BridgeEmulator/openssl.conf /opt/hue-emulator/
+    chmod +x /opt/hue-emulator/genCert.sh
+else
+    echo "ui update"
+fi
 
-#cd /
 if [ -d diyhueUI ]; then
- rm -r diyhueUI
+    rm -r diyhueUI
 fi
 mkdir diyhueUI
 curl -sL https://github.com/diyhue/diyHueUI/releases/latest/download/DiyHueUI-release.zip -o diyHueUI.zip
+#curl -sL https://github.com/hendriksen-mark/diyHueUI/releases/latest/download/DiyHueUI-release.zip -o diyHueUI.zip
 unzip -qo diyHueUI.zip -d diyhueUI
 rm diyHueUI.zip
-#cd diyhueUI
 cp -r diyhueUI/index.html /opt/hue-emulator/flaskUI/templates/
 cp -r diyhueUI/static /opt/hue-emulator/flaskUI/
 
-curl -s localhost/restart
+curl -s $1/restart
